bot_backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
Django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
whitenoise>=6.6

# ── Database ──────────────────────────────────────────────────────────────────
psycopg2-binary>=2.9

# ── Cache ─────────────────────────────────────────────────────────────────────
# Using Django's built-in LocMemCache — no Redis needed

# ── AI / LLM ──────────────────────────────────────────────────────────────────
openai>=1.30          # used for both Groq (OpenAI-compatible) and native OpenAI

# ── Crawling / parsing ────────────────────────────────────────────────────────
requests>=2.31
certifi>=2024.0
beautifulsoup4>=4.12
playwright>=1.40          # optional: auto-fallback to requests if unavailable

# ── Keyword extraction ────────────────────────────────────────────────────────
keybert>=0.8

# ── Content generation extras ────────────────────────────────────────────────
youtube-transcript-api>=0.6   # optional: YouTube article generation
Pillow>=10.0                  # optional: image handling

# ── Config ────────────────────────────────────────────────────────────────────
python-dotenv>=1.0
